data-pipeline-docker/requirements.txt
pandas
kafka-python
sqlalchemy
pymysql
pymongo
python-dotenv